refactor(docs): move chrome into layout route
