+++ src/lib/docs-page-data.ts
@@ -1,5 +1,5 @@
 import { notFound } from "@tanstack/react-router";
-import { getDocsNavTree, getDocsTocItems } from "./docs-navigation";
+import { getDocsTocItems } from "./docs-navigation";
 import { source } from "./source";
 
 export function getDocPage(slugs: Array<string>) {
@@ -15,6 +15,5 @@ export function getDocPage(slugs: Array<string>) {
     title: page.data.title,
     description: page.data.description,
     toc: getDocsTocItems(page.data.toc),
-    navTree: getDocsNavTree(),
   };
 }
